Extract receipt logic to a super class
